Refactor bt-caching stuff to it's own module
